fix: moved CommandName enum in types
