--- src/cli/commands/constants.ts
@@ -1,6 +1,0 @@
-export enum CommandName {
-  VERSION = 'version',
-  HELP = 'help',
-  IMPORT = 'import',
-  GENERATE = 'generate',
-}
